backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
pydantic==1.10.9
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
requests==2.32.3
celery==5.3.6
redis==5.0.1
supabase==2.3.1
openai==1.12.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
httpx>=0.24.0,<0.26.0
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.2.0
python-slugify==8.0.4
tweepy==4.14.0
instagrapi==2.0.0
google-api-python-client==2.108.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9 